Problem : Duplicated code with sse connection and get asset, get alert list
