+++ src/web/src/alert_list.ecpp
@@ -43,6 +43,7 @@
 #include "db/assets.h"
 #include "cleanup.h"
 #include "helpers.h"
+#include "utils_json.h"
 
 #define RFC_ALERTS_LIST "rfc-alerts-list"
 
@@ -364,54 +365,20 @@ if (streq (part, "LIST")) {
 %       frame = zmsg_pop (recv_msg);
 %       continue;
 %   }
-%   char buff[64];
-%   rv = calendar_to_datetime (fty_proto_time (decoded), buff, 64);
-%   if (rv == -1) {
-%       log_error ("can't convert %" PRIu64 "to calendar time, skipping element '%s'", fty_proto_time (decoded), fty_proto_rule (decoded) );
-%       fty_proto_destroy (&decoded);
-%       frame = zmsg_pop (recv_msg);
+%   
+%   std::string jsonAlert = getJsonAlert(connection,decoded);
+%   
+%   if (jsonAlert.empty()){
 %       continue;
 %   }
-%   log_debug ("persist::select_asset_element_web_byName ('%s')", fty_proto_name (decoded));
-%
-%   auto asset_element = persist::select_asset_element_web_byName (connection, fty_proto_name (decoded));
-%   if (asset_element.status != 1) {
-%       log_error ("%s", asset_element.msg.c_str ());
-%       fty_proto_destroy (&decoded);
-%       frame = zmsg_pop (recv_msg);
-%       continue;
-%   }
-%
-%   std::pair<std::string,std::string> asset_element_names = persist::id_to_name_ext_name (asset_element.item.id);
-%   if (asset_element_names.first.empty () && asset_element_names.second.empty ())
-%       http_die ("internal-error", "Database error");
 %
 %   if (first) {
-    {
-        <$$ utils::json::jsonify ("timestamp", buff) $>,
-        <$$ utils::json::jsonify ("rule_name", fty_proto_rule (decoded)) $>,
-        <$$ utils::json::jsonify ("element_id",  fty_proto_name (decoded)) $>,
-        <$$ utils::json::jsonify ("element_name", asset_element_names.second) $>,
-        <$$ utils::json::jsonify ("element_type", asset_element.item.type_name) $>,
-        <$$ utils::json::jsonify ("element_sub_type", utils::strip (asset_element.item.subtype_name)) $>,
-        <$$ utils::json::jsonify ("state", fty_proto_state (decoded)) $>,
-        <$$ utils::json::jsonify ("severity", fty_proto_severity (decoded)) $>,
-        <$$ utils::json::jsonify ("description", fty_proto_description (decoded)) $>
-    }
-%     first = false;
+      <$$ jsonAlert  $>
+%      first = false;
 %   }
 %   else {
-, {
-        <$$ utils::json::jsonify ("timestamp", buff) $>,
-        <$$ utils::json::jsonify ("rule_name", fty_proto_rule (decoded)) $>,
-        <$$ utils::json::jsonify ("element_id", fty_proto_name (decoded)) $>,
-        <$$ utils::json::jsonify ("element_name", asset_element_names.second) $>,
-        <$$ utils::json::jsonify ("element_type", asset_element.item.type_name) $>,
-        <$$ utils::json::jsonify ("element_sub_type", utils::strip (asset_element.item.subtype_name)) $>,
-        <$$ utils::json::jsonify ("state", fty_proto_state (decoded)) $>,
-        <$$ utils::json::jsonify ("severity", fty_proto_severity (decoded)) $>,
-        <$$ utils::json::jsonify ("description", fty_proto_description (decoded)) $>
-}
+, 
+      <$$ jsonAlert  $>
 %   }
 %   fty_proto_destroy (&decoded);
 %   frame = zmsg_pop (recv_msg);
